refactor(charts): extract shared utilities into _shared module
Deduplicates formatValue, getNumericValue, calculateNiceTicks,
getGridDasharray, and useContainerDimensions across all 10 chart
components into src/components/charts/_shared/.

Removes ~560 lines of duplicated code. Fixes now only need to be
applied once.

Closes #33
