+++ src/components/charts/_shared/hooks.ts
@@ -1,0 +1,28 @@
+"use client";
+
+import { useRef, useState } from "react";
+import { useIsomorphicLayoutEffect } from "../../../../lib/hooks";
+
+export function useContainerDimensions() {
+  const ref = useRef<HTMLDivElement>(null);
+  const [width, setWidth] = useState(0);
+
+  useIsomorphicLayoutEffect(() => {
+    const el = ref.current;
+    if (!el) return;
+    let rafId = 0;
+    const update = () => {
+      cancelAnimationFrame(rafId);
+      rafId = requestAnimationFrame(() => setWidth(el.getBoundingClientRect().width));
+    };
+    update();
+    const ro = new ResizeObserver(update);
+    ro.observe(el);
+    return () => {
+      cancelAnimationFrame(rafId);
+      ro.disconnect();
+    };
+  }, []);
+
+  return [ref, width] as const;
+}
