+++ src/components/charts/_shared/types.ts
@@ -1,0 +1,1 @@
+export type ChartDataItem = Record<string, unknown>;
