+++ src/components/charts/bar-chart/index.tsx
@@ -1,13 +1,11 @@
 "use client";
 
 import * as React from "react";
-import { memo, useMemo, useState, useRef, useCallback } from "react";
-import { useIsomorphicLayoutEffect } from "../../../../lib/hooks";
+import { memo, useMemo, useState, useCallback } from "react";
 import { motion, AnimatePresence, useReducedMotion } from "framer-motion";
 import { cn } from "../../../../lib/utils";
-
-// Types
-type ChartDataItem = Record<string, unknown>;
+import { formatValue, getNumericValue, calculateNiceTicks, getGridDasharray, useContainerDimensions } from "../_shared";
+import type { ChartDataItem } from "../_shared";
 
 interface BarChartProps<T extends ChartDataItem> {
   readonly data: readonly T[];
@@ -36,86 +34,6 @@ const DEFAULT_HEIGHT = 300;
 const MARGIN = { top: 20, right: 20, bottom: 40, left: 50 };
 const ANIMATION_EASING = [0.4, 0, 0.2, 1] as const;
 const HOVER_DURATION = 0.2;
-
-// Utilities
-function formatValue(value: unknown): string {
-  if (typeof value === 'number') {
-    if (Math.abs(value) >= 1000000) {
-      return `${(value / 1000000).toFixed(1)}M`;
-    } else if (Math.abs(value) >= 1000) {
-      return `${(value / 1000).toFixed(1)}K`;
-    }
-    return value.toLocaleString();
-  }
-  return String(value);
-}
-
-function getNumericValue(data: ChartDataItem, key: keyof ChartDataItem): number {
-  const value = data[key];
-  if (typeof value === 'number' && isFinite(value)) return value;
-  if (typeof value === 'string') {
-    const parsed = parseFloat(value.replace(/[,$%\s]/g, ''));
-    if (isFinite(parsed)) return parsed;
-    if (process.env.NODE_ENV === 'development') {
-      console.warn(`[BarChart] Could not parse value "${value}" for key "${String(key)}". Using 0.`);
-    }
-  }
-  if (process.env.NODE_ENV === 'development' && value !== null && value !== undefined) {
-    console.warn(`[BarChart] Unexpected value type for key "${String(key)}": ${typeof value}. Using 0.`);
-  }
-  return 0;
-}
-
-function calculateNiceTicks(min: number, max: number, count = 5): number[] {
-  if (min === max) return [min];
-  const range = max - min;
-  const roughStep = range / (count - 1);
-  const magnitude = Math.pow(10, Math.floor(Math.log10(roughStep)));
-  const normalizedStep = roughStep / magnitude;
-  let niceStep: number;
-  if (normalizedStep <= 1) niceStep = magnitude;
-  else if (normalizedStep <= 2) niceStep = 2 * magnitude;
-  else if (normalizedStep <= 5) niceStep = 5 * magnitude;
-  else niceStep = 10 * magnitude;
-  const niceMin = Math.floor(min / niceStep) * niceStep;
-  const niceMax = Math.ceil(max / niceStep) * niceStep;
-  const ticks: number[] = [];
-  for (let tick = niceMin; tick <= niceMax; tick += niceStep) {
-    ticks.push(tick);
-  }
-  return ticks;
-}
-
-function getGridDasharray(gridStyle: 'solid' | 'dashed' | 'dotted'): string {
-  switch (gridStyle) {
-    case 'solid': return 'none';
-    case 'dotted': return '2 4';
-    case 'dashed':
-    default: return '4 4';
-  }
-}
-
-function useContainerDimensions() {
-  const ref = useRef<HTMLDivElement>(null);
-  const [width, setWidth] = useState(0);
-
-  useIsomorphicLayoutEffect(() => {
-    const element = ref.current;
-    if (!element) return;
-
-    const updateWidth = () => {
-      setWidth(element.getBoundingClientRect().width);
-    };
-
-    updateWidth();
-    const resizeObserver = new ResizeObserver(updateWidth);
-    resizeObserver.observe(element);
-
-    return () => resizeObserver.disconnect();
-  }, []);
-
-  return [ref, width] as const;
-}
 
 // Loading/Error States
 function LoadingState({
