+++ src/components/charts/funnel-chart/index.tsx
@@ -3,11 +3,8 @@
 import * as React from "react";
 import { memo, useMemo, useCallback } from "react";
 import { motion, AnimatePresence, useReducedMotion } from "framer-motion";
-import { useIsomorphicLayoutEffect } from "../../../../lib/hooks";
 import { cn } from "../../../../lib/utils";
-
-// Types
-type ChartDataItem = Record<string, unknown>;
+import { formatValue, getNumericValue, useContainerDimensions, type ChartDataItem } from "../_shared";
 
 interface FunnelChartProps<T extends ChartDataItem> {
   readonly data: readonly T[];
@@ -78,54 +75,12 @@ const H_ROW_GAP = 8;      // horizontal: gap between bars
 const H_CONV_H = 16;      // horizontal: conversion rate row height
 
 // Utilities
-function getNumericValue(data: ChartDataItem, key: keyof ChartDataItem): number {
-  const val = data[key];
-  if (typeof val === "number" && isFinite(val)) return val;
-  if (typeof val === "string") {
-    const parsed = parseFloat(val.replace(/[,$%\s]/g, ""));
-    if (isFinite(parsed)) return parsed;
-  }
-  return 0;
-}
-
-function formatValue(value: number): string {
-  if (Math.abs(value) >= 1_000_000) return `${(value / 1_000_000).toFixed(1)}M`;
-  if (Math.abs(value) >= 1_000) return `${(value / 1_000).toFixed(1)}K`;
-  return value.toLocaleString();
-}
-
 function buildPolygon(cx: number, topW: number, bottomW: number, y: number, h: number): string {
   const tl = cx - topW / 2;
   const tr = cx + topW / 2;
   const bl = cx - bottomW / 2;
   const br = cx + bottomW / 2;
   return `${tl},${y} ${tr},${y} ${br},${y + h} ${bl},${y + h}`;
-}
-
-import { useRef } from "react";
-
-function useContainerDimensions() {
-  const ref = useRef<HTMLDivElement>(null);
-  const [width, setWidth] = React.useState(0);
-
-  useIsomorphicLayoutEffect(() => {
-    const el = ref.current;
-    if (!el) return;
-    let rafId = 0;
-    const update = () => {
-      cancelAnimationFrame(rafId);
-      rafId = requestAnimationFrame(() => setWidth(el.getBoundingClientRect().width));
-    };
-    update();
-    const ro = new ResizeObserver(update);
-    ro.observe(el);
-    return () => {
-      cancelAnimationFrame(rafId);
-      ro.disconnect();
-    };
-  }, []);
-
-  return [ref, width] as const;
 }
 
 // States
